Update dependencies and refactor workout handling to improve event publishing and data types
- Updated springdoc-openapi version to 2.8.6 in pom.xml.
- Enhanced SecurityConfig to include additional public endpoints for Swagger and OpenAPI.
- Introduced WorkoutEvent for better event handling in WorkoutController.
- Refactored Workout and WorkoutUploadDto to use double for distance instead of BigDecimal.
- Added aggregateForGoalPeriod query in WorkoutRepository for goal tracking.
- Updated WorkoutEventListener to refresh user summaries and update goals upon workout events.
- Adjusted tests to reflect changes in data types for distance.

+++ src/main/java/com/promax/workout/dto/WorkoutUploadDto.java
@@ -27,7 +27,7 @@ public class WorkoutUploadDto {
 
     @NotNull(message = "Distance cannot be empty")
     @DecimalMin(value = "0.0", message = "Distance cannot be negative")
-    private BigDecimal distanceKm;
+    private double distanceKm;
 
     @NotNull(message = "Calories burned cannot be empty")
     @Min(value = 0, message = "Calories burned cannot be negative")
@@ -41,7 +41,7 @@ public class WorkoutUploadDto {
     }
 
     // Constructor with parameters
-    public WorkoutUploadDto(WorkoutType workoutType, Integer durationMinutes, BigDecimal distanceKm,
+    public WorkoutUploadDto(WorkoutType workoutType, Integer durationMinutes, double distanceKm,
             Integer caloriesBurned, String notes) {
         this.workoutType = workoutType;
         this.durationMinutes = durationMinutes;
@@ -67,11 +67,11 @@ public class WorkoutUploadDto {
         this.durationMinutes = durationMinutes;
     }
 
-    public BigDecimal getDistanceKm() {
+    public double getDistanceKm() {
         return distanceKm;
     }
 
-    public void setDistanceKm(BigDecimal distanceKm) {
+    public void setDistanceKm(double distanceKm) {
         this.distanceKm = distanceKm;
     }
